+++ components/OauthLogin/index.tsx
@@ -1,31 +1,25 @@
 'use client'
 
-import { Facebook, Github } from "lucide-react"
 import { signIn } from "next-auth/react"
 
 import React from 'react'
-import { Google } from "../shared/icons"
+import { Github, Google } from "../shared/icons"
+import FacebookIcon from "../shared/icons/facebook"
 
 function OauthLogin() {
-  return (
-    <div className="flex justify-center items-center gap-4">
-                    <button onClick={()=>signIn("google")}>
-                        <div>
-                            <Google />
-                        </div>
-                    </button>
-                    <button onClick={()=>signIn("facebook")}>
-                        <div>
-                            <Facebook />
-                        </div>
-                    </button>
-                    <button onClick={()=>signIn("github")}>
-                        <div>
-                            <Github/>
-                        </div>
-                    </button>
-                </div>
-  )
+    return (
+        <div className="flex justify-around items-center gap-4">
+            <button className="w-20 h-20 flex justify-center items-center" onClick={() => signIn("google")}>
+                <Google />
+            </button>
+            <button className="w-20 h-20 flex justify-center items-center " onClick={() => signIn("facebook")}>
+                <FacebookIcon />
+            </button>
+            <button className="w-20 h-20 flex justify-center items-center " onClick={() => signIn("github")}>
+                <Github />
+            </button>
+        </div>
+    )
 }
 
 export default OauthLogin
