+++ crud/user.ts
@@ -5,6 +5,8 @@ import { CreateUserDTO, CredentialAuthDTO } from "./DTOs";
 import bcrypt from "bcrypt"
 import { AdapterAccount } from "next-auth/adapters";
 import { randomUUID } from "crypto";
+import {verify} from 'jsonwebtoken';
+import { sendPasswordEmail } from "@/lib/externalRequests/sendgrid";
 
 async function createWithPassword(user: CreateUserDTO, prismaClient: PrismaClient) {
     const users = prismaClient.user;
@@ -78,6 +80,24 @@ async function update(userId: string, user: CreateUserDTO, prismaClient: PrismaC
         });
         return updatedUser
     }
+
+}
+
+
+export async function reset(token: string, password: string, prismaClient: PrismaClient) {
+    const users = prismaClient.user;
+    const { email } = verify(token as string, process.env.NEXTAUTH_SECRET as string) as { email: string };
+
+    const hashedPassword = await bcrypt.hash(password, 10)
+    const updated = await users.update({
+        where: { email }, data: {
+
+            password: hashedPassword
+        }
+    });
+    await sendPasswordEmail({ email, password, subject: "New Cyberoni credentials" })
+    return true;
+
 
 }
 async function remove(userId: string, prismaClient: PrismaClient) {
